requirements.txt
python-jobspy==1.1.80
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
notion-client==2.2.1
python-dotenv==1.0.1
